backend/requirements-dev.txt
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
sqlalchemy-seeder==0.3.0
PyYAML==6.0.1
